Removed test file, separated actual test code from implmentation, and added method comment
