--- test.py
@@ -1,6 +1,0 @@
-def pass_string(s):
-    return s
-
-
-def test_pass():
-    assert pass_string("s") == "s"
